initial attempt at updating the design to 128GB
things that are odd:
- not sure what "CDAT DSMAS (DPA length)" update corresponds to in
cafu_common_pkg.sv. The doc on 04/07/25 says it should be updated from
0x4 to 0x20. However, it doesn't look like this field exists in
cafu_common_pkg.sv. An older version of the doc (released 12/10/24) says
we should update CDAT DSMAS_5 from 0x4 to 0x20, and this is what this
commit does.

- not sure what the CDAT CDAT_1 checksum byte should be updated to. our
  design example generated from quartus 25.1 shows that the checksum
value is 0x55 (for 16GB, consistent with the 12/10/24 doc). however,
the doc from 04/07/25 says it should be 0x44. this commit follows the
newer doc and updates the value to 0x28; however, the old doc suggests
the checksum should be updated to 0x39.